requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
ffmpeg==1.4
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
urllib3==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
ffmpeg==1.4
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
pydub==0.25.1
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
urllib3==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
ffmpeg==1.4
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
pydub==0.25.1
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sk-video==1.1.10
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
urllib3==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
ffmpeg==1.4
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.5.0
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sk-video==1.1.10
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
urllib3==2.0.2
gradio